+++ _Typings/editorApi/types/transform.d.ts
@@ -1,0 +1,7 @@
+declare module "editor:type" {
+    type Transform = {
+        Position: Float3
+        Rotation: Float3
+        Scale: Float3
+    }
+}