+++ _Typings/gameApi/alias/item.d.ts
@@ -1,5 +1,0 @@
-declare module "game:alias" {
-    import { Item as ItemAlias } from "game:type"
-
-    type Item = ItemAlias
-}