+++ _Typings/gameApi/alias/utils.d.ts
@@ -1,0 +1,3 @@
+declare module "game:alias" {
+    export { BallType, CollisionEvent, EaseType, Events, Item, RaycastResult, Tween } from "game:type"
+}